--- sedov_blast_wave/sedov_update.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "../wash_mockapi.hpp"
-#include "sedov_box.hpp"
-#include "sedov_consts.hpp"
-
-void update_kernel(wash::Particle& i);
