+++ src/examples/sedov_blast_wave/force.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "../wash_mockapi.hpp"
-#include "sedov_box.hpp"
-#include "sedov_consts.hpp"
+#include "../../../wash_mockapi.hpp"
+#include "box.hpp"
+#include "consts.hpp"
 
 void force_kernel(wash::Particle& i, const std::vector<wash::Particle>& neighbours);
