+++ src/examples/sedov_blast_wave/init.hpp
@@ -3,7 +3,7 @@
 #define _USE_MATH_DEFINES
 #include <cmath>
 
-#include "../wash_mockapi.hpp"
-#include "sedov_consts.hpp"
+#include "../../../wash_mockapi.hpp"
+#include "consts.hpp"
 
 void init();
