+++ src/examples/sedov_blast_wave/update.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../../../wash_mockapi.hpp"
+#include "box.hpp"
+#include "consts.hpp"
+
+void update_kernel(wash::Particle& i);
